desktop-agent/requirements.txt
mss==9.0.1
pyperclip==1.8.2
pynput==1.7.6
requests==2.31.0
python-dotenv==1.0.0
pillow==10.0.1
pystray==0.19.5
colorama==0.4.6
cryptography==41.0.3
psutil==5.9.6
pyaudio==0.2.13
pyinstaller==6.0.0
numpy==1.24.3
